services/vehicle-runtime/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
requests==2.32.3
numpy==2.1.3
Pillow==11.0.0
onnxruntime==1.20.1
opencv-python-headless==4.10.0.84
pyserial==3.5
pytest==8.3.4
